Make t0..t3 equal in size to a register. Fixes compilation with llvm-g++ on x86_64.
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@10024 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/video/Simple3xScaler.cc
@@ -345,7 +345,7 @@ void Blur_1on3<Pixel>::operator()(const Pixel* in, Pixel* out, unsigned long dst
 		Blur_1on3_4_SSE(in, out, dstWidth, &c);
 		return;
 	#else
-		unsigned t0, t1, t2, t3;
+		unsigned long t0, t1, t2, t3;
 		asm volatile (
 			"pxor      %%mm0, %%mm0;"
 			"pshufw    $0,40(%[CNST]),%%mm1;"
